scripts/pyproject.toml
[project]
name = "description-rewrite"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
]
